move fauxhdfbuilder.py from bapsflib._hdf.maps.tests to bapsflib.faux
