+++ bapsflib/faux/fauxhdfbuilder.py
